:truck: Move cms assemblies to assemblies/cms subfolder
