+++ frontend/src/components/assemblies/cms/Title.js
@@ -1,16 +1,16 @@
 import { func, object } from 'prop-types';
-import Input from '../form/Input';
-import Label from '../form/Label';
+import Input from '../../form/Input';
+import Label from '../../form/Label';
 
 export default function Title({ component, onChange }) {
   return (
     <>
-      <Label>{component.type === 'title' ? 'title' : 'subtitle'}</Label>
+      <Label>{component.type}</Label>
       <Input
         align="left"
         id="content"
         type="text"
-        placeholder={component.type === 'title' ? 'Enter a title...' : 'Enter a subtitle...'}
+        placeholder={component.type === 'subtitle' ? 'Enter a subtitle...' : 'Enter a title...'}
         value={component.content || ''}
         onChange={onChange} />
     </>
